Simplifying Hlaalu Scrip Exchange Scripts
Deleted the global script, exchange now happens directly from the NPC script. Makes the NPC scripts quite a bit longer but ultimately removing the interplay between these two scripts and putting it all in a single script is the right thing to do as it eliminates the need for most frame delays and simplifies the procedure.