edit sky_ex_nord_df_02_08.nif for Markarth
Previous versions of this doorframe had a decorative wooden arch, which was one-sided, so I deleted the "back" of the frame. It turns out that the mesh was used backwards dozens of times in Markarth, with the one-sided arch deliberately buried in the walls of buildings. Per a conversation with Tristior, this version restores the back (now the front) and deletes the unused arch entirely. https://discord.com/channels/261789910594355200/391968847969648663/1361340136662503525